+++ hpc/nvidia/cuda/fakeCUDA/frontend/hello.c
@@ -1,0 +1,8 @@
+__attribute__((device)) void kernel1(void)
+{
+}
+
+int main(void)
+{
+	return 0;
+}